Server/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
debugpy==1.8.12
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
urllib3==2.3.0
